Larger query evaluation performance fix
Fixed the state validation logic for ongoing queries with larger output sizes being significantly slowed down due to the new state validation logic introduced with filter support: these are now only executed in a debug implementation of the ongoing query type, which should not be instantiated when measuring performance
