+++ sparql/src/commonMain/kotlin/dev/tesserakt/sparql/OngoingQueryEvaluation.kt
@@ -2,76 +2,19 @@ package dev.tesserakt.sparql
 
 import dev.tesserakt.rdf.types.MutableStore
 import dev.tesserakt.rdf.types.Quad
-import dev.tesserakt.sparql.runtime.evaluation.DataAddition
-import dev.tesserakt.sparql.runtime.evaluation.DataDeletion
-import dev.tesserakt.sparql.runtime.query.QueryState
-import dev.tesserakt.util.replace
 
+interface OngoingQueryEvaluation<RT> {
 
-class OngoingQueryEvaluation<RT>(private val query: QueryState<RT, *>) {
+    val results: Collection<RT>
 
-    private val _results = mutableMapOf<RT, Int>()
-    val results get() = _results.flatMap { entry -> List(entry.value) { entry.key } }
+    fun subscribe(store: MutableStore)
 
-    private val processor = query.Processor()
+    fun unsubscribe(store: MutableStore)
 
-    private val listener = object: MutableStore.Listener {
-        override fun onQuadAdded(quad: Quad) {
-            add(quad)
-        }
+    fun add(quad: Quad)
 
-        override fun onQuadRemoved(quad: Quad) {
-            remove(quad)
-        }
-    }
+    fun remove(quad: Quad)
 
-    init {
-        // setting the query state as the current results
-        processor.state().forEach {
-            _results[it] = 1
-        }
-    }
-
-    fun subscribe(store: MutableStore) {
-        store.forEach { quad ->
-            processor.process(DataAddition(quad)).forEach { process(it) }
-        }
-        store.addListener(listener)
-    }
-
-    fun unsubscribe(store: MutableStore) {
-        store.removeListener(listener)
-    }
-
-    fun add(quad: Quad) {
-        processor.process(DataAddition(quad)).forEach { process(it) }
-        ensureValidState()
-    }
-
-    fun remove(quad: Quad) {
-        processor.process(DataDeletion(quad)).forEach { process(it) }
-        ensureValidState()
-    }
-
-    fun debugInformation(): String {
-        return processor.debugInformation()
-    }
-
-    private fun process(change: QueryState.ResultChange<Bindings>) {
-        when (val mapped = query.process(change)) {
-            is QueryState.ResultChange.New<*> -> {
-                _results.replace(mapped.value) { current -> (current ?: 0) + 1 }
-            }
-            is QueryState.ResultChange.Removed<*> -> {
-                _results.replace(mapped.value) { current -> (current ?: 0) - 1 }
-            }
-        }
-    }
-
-    private fun ensureValidState() {
-        _results.forEach {
-            check(it.value >= 0) { "${it.key} was not removed properly from the result list as it did not exist!" }
-        }
-    }
+    fun debugInformation(): String
 
 }
